requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
transformers==4.35.2
torch==2.7.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
beautifulsoup4==4.12.2
requests==2.31.0
nltk==3.8.1
python-multipart==0.0.6